refactor(pdf): extract batch_report.rs and comparison_report.rs from pdf/mod.rs
Moves generate_batch_pdf + 6 batch helpers → batch_report.rs,
generate_comparison_pdf → comparison_report.rs.
pdf/mod.rs reduced from 3625 to 1981 lines (1231 non-test lines).
No logic changes; generate_pdf stays in mod.rs with its helpers and tests.
