+++ src/output/pdf/mod.rs
@@ -4,6 +4,8 @@
 //! directly to renderreport components. Zero data transformation here.
 
 mod batch;
+mod batch_report;
+mod comparison_report;
 mod cover;
 mod design;
 mod detail_modules;
@@ -12,10 +14,13 @@ mod helpers;
 mod history;
 mod modules;
 
+pub use self::batch_report::generate_batch_pdf;
+pub use self::comparison_report::generate_comparison_pdf;
+
 use renderreport::components::advanced::{
     ChecklistPanel, ChecklistRow, DevicePreview, DiagnosisPanel, DiagnosisRow,
-    DominantIssueSpotlight, ImpactGrid, ImpactGridCard, KeyValueList, List, MetricStrip,
-    MetricStripItem, PageBreak, PhaseBlock, SectionHeaderSplit, TableOfContents,
+    DominantIssueSpotlight, ImpactGrid, ImpactGridCard, List, MetricStrip, MetricStripItem,
+    PageBreak, PhaseBlock, SectionHeaderSplit, TableOfContents,
 };
 use renderreport::components::text::{Label, TextBlock};
 use renderreport::components::TagCloud;
@@ -23,35 +28,23 @@ use renderreport::prelude::Image;
 use renderreport::prelude::*;
 
 // Composite components
-use renderreport::components::{
-    AuditTable, BenchmarkRow, BenchmarkTable, ComparisonModule, ModuleComparison, SeverityOverview,
-    SummaryBox, TableColumn,
-};
+use renderreport::components::{AuditTable, SeverityOverview, SummaryBox, TableColumn};
 
-use crate::audit::{normalize, AuditReport, BatchReport};
-use crate::cli::ReportLevel;
-use crate::i18n::I18n;
-use crate::output::builder::{build_batch_presentation, build_view_model};
-use crate::output::report_model::*;
-use crate::util::truncate_url;
-
-use self::batch::build_batch_overview_grid;
-use self::cover::{
-    auditmysite_wordmark_path, batch_certificate_label, build_batch_cover_score_row,
-    build_cover_score_row, certificate_badge_path,
-};
+use self::cover::{auditmysite_wordmark_path, build_cover_score_row, certificate_badge_path};
 use self::detail_modules::{
     render_ai_visibility, render_budget_violations, render_content_visibility, render_dark_mode,
     render_journey, render_mobile, render_performance, render_security, render_seo,
     render_source_quality, render_tech_stack, render_ux,
 };
 use self::findings::{first_sentence, render_finding_technical, render_key_finding_block};
-use self::helpers::{
-    component_json, create_engine, extract_domain, priority_label_i18n, role_label_i18n,
-    severity_label_i18n, soft_flow_group,
-};
+use self::helpers::{component_json, create_engine, extract_domain, soft_flow_group};
 use self::history::{render_history_section, render_methodology_section};
 use self::modules::{build_summary_overview, build_was_jetzt_tun_table, WasJetztTunContent};
+use crate::audit::{normalize, AuditReport};
+use crate::cli::ReportLevel;
+use crate::i18n::I18n;
+use crate::output::builder::build_view_model;
+use crate::output::report_model::*;
 
 const WORDMARK_ASSET: &str = "/auditmysite-wordmark.svg";
 const CUSTOM_COVER_LOGO_ASSET: &str = "/cover-logo-custom";
@@ -1012,14 +1005,14 @@ fn build_raw_audit_snapshot(vm: &ReportViewModel, i18n: &I18n) -> SummaryBox {
         })
 }
 
-fn cover_logo_asset(config: &ReportConfig) -> &'static str {
+pub(super) fn cover_logo_asset(config: &ReportConfig) -> &'static str {
     match config.logo_path.as_ref() {
         Some(path) if path.exists() => CUSTOM_COVER_LOGO_ASSET,
         _ => WORDMARK_ASSET,
     }
 }
 
-fn register_cover_logo_asset(
+pub(super) fn register_cover_logo_asset(
     mut builder: renderreport::engine::ReportBuilder,
     config: &ReportConfig,
     cover_logo_asset: &'static str,
@@ -1228,1648 +1221,6 @@ fn business_relevance(page_type: Option<&str>, url: &str, locale: &str) -> &'sta
         Some("Thin / Minimal Content") => low,
         _ => medium,
     }
-}
-
-// ─── Helper: Batch Report Assessment & Key Points ──────────────────────────
-
-/// Clear batch assessment — no score, just interpretation
-fn build_batch_assessment(
-    summary: &crate::output::report_model::PortfolioSummary,
-    dist: &SeverityDistribution,
-    i18n: &I18n,
-) -> String {
-    let en = i18n.locale() == "en";
-    let score = summary.average_score.round() as u32;
-    if dist.critical > 0 && score < 50 {
-        if en {
-            "Critical barriers — not WCAG conformant".to_string()
-        } else {
-            "Kritische Barrieren — nicht WCAG-konform".to_string()
-        }
-    } else if dist.critical > 0 {
-        if en {
-            "Technically solid, but legally risky".to_string()
-        } else {
-            "Technisch solide, aber rechtlich riskant".to_string()
-        }
-    } else if dist.high > 0 {
-        if en {
-            "Good foundation, but not accessible".to_string()
-        } else {
-            "Gute Basis, aber nicht barrierefrei".to_string()
-        }
-    } else if score >= 85 {
-        if en {
-            "Largely accessible — polish".to_string()
-        } else {
-            "Weitgehend barrierefrei — Feinschliff".to_string()
-        }
-    } else if en {
-        "Solid foundation with room to optimize".to_string()
-    } else {
-        "Solide Grundlage mit Optimierungspotenzial".to_string()
-    }
-}
-
-/// 3 key takeaways for batch report
-fn build_batch_key_points(
-    pres: &BatchPresentation,
-    dist: &SeverityDistribution,
-    i18n: &I18n,
-) -> Vec<String> {
-    let en = i18n.locale() == "en";
-    let mut points = Vec::with_capacity(3);
-
-    // Point 1: Critical/high count across all URLs
-    let ch = dist.critical + dist.high;
-    if ch > 0 {
-        if en {
-            points.push(format!(
-                "{} critical/high violations across {} URLs",
-                ch, pres.portfolio_summary.total_urls
-            ));
-        } else {
-            points.push(format!(
-                "{} kritische/hohe Verstöße über {} URLs hinweg",
-                ch, pres.portfolio_summary.total_urls
-            ));
-        }
-    }
-
-    // Point 2: Dominant/recurring issue
-    if let Some(top) = pres.top_issues.first() {
-        if en {
-            points.push(format!(
-                "Main issue: {} ({} occurrences on {} URLs)",
-                top.title,
-                top.occurrence_count,
-                top.affected_urls.len()
-            ));
-        } else {
-            points.push(format!(
-                "Hauptproblem: {} ({} Vorkommen auf {} URLs)",
-                top.title,
-                top.occurrence_count,
-                top.affected_urls.len()
-            ));
-        }
-    }
-
-    // Point 3: Legal status
-    if dist.critical > 0 {
-        if en {
-            points.push(
-                "WCAG Level A violations detected automatically — manual review needed for a defensible BFSG classification".to_string(),
-            );
-        } else {
-            points.push(
-                "WCAG-Level-A-Verstöße automatisiert erkannt — manuelle Prüfung für belastbare BFSG-Einordnung nötig".to_string(),
-            );
-        }
-    } else if dist.high > 0 {
-        if en {
-            points.push(
-                "No Level A violations, but structural weaknesses on multiple pages".to_string(),
-            );
-        } else {
-            points.push(
-                "Keine Level-A-Verstöße, aber strukturelle Schwächen auf mehreren Seiten"
-                    .to_string(),
-            );
-        }
-    } else if en {
-        points.push(
-            "No automatically detectable critical barriers — manual review recommended."
-                .to_string(),
-        );
-    } else {
-        points.push(
-            "Keine automatisiert erkennbaren kritischen Barrieren — manuelle Prüfung empfohlen."
-                .to_string(),
-        );
-    }
-
-    points
-}
-
-/// Concrete quick actions for batch report
-fn build_batch_quick_actions(pres: &BatchPresentation, i18n: &I18n) -> Vec<(String, String)> {
-    let en = i18n.locale() == "en";
-    let timeframe_label = |effort: Effort| -> &'static str {
-        match (effort, en) {
-            (Effort::Quick, true) => "1–2 days",
-            (Effort::Quick, false) => "1–2 Tage",
-            (Effort::Medium, true) => "3–5 days",
-            (Effort::Medium, false) => "3–5 Tage",
-            (Effort::Structural, true) => "1–2 weeks",
-            (Effort::Structural, false) => "1–2 Wochen",
-        }
-    };
-
-    let mut actions: Vec<(String, String)> = Vec::new();
-
-    for item in &pres.action_plan.quick_wins {
-        if actions.len() >= 3 {
-            break;
-        }
-        let action_lower = item.action.to_lowercase();
-        let scope = if action_lower.contains("alle") || action_lower.contains("global") {
-            " (global)"
-        } else {
-            ""
-        };
-        actions.push((
-            format!("{}{}", item.action, scope),
-            timeframe_label(item.effort).to_string(),
-        ));
-    }
-
-    // Fallback from top issues
-    if actions.is_empty() {
-        for group in pres.top_issues.iter().take(3) {
-            actions.push((
-                first_sentence(&group.recommendation).to_string(),
-                timeframe_label(group.effort).to_string(),
-            ));
-        }
-    }
-
-    actions
-}
-
-/// Enhanced action plan with effort + scope columns
-fn render_batch_action_plan_enhanced(
-    mut builder: renderreport::engine::ReportBuilder,
-    plan: &ActionPlan,
-    i18n: &I18n,
-) -> renderreport::engine::ReportBuilder {
-    let en = i18n.locale() == "en";
-    let effort_col = if en { "Effort" } else { "Aufwand" };
-    let role_col = if en { "Role" } else { "Rolle" };
-    let scope_global = "global";
-    let scope_content = "Content";
-    let scope_component = if en { "Component" } else { "Komponente" };
-
-    let render_section = |mut b: renderreport::engine::ReportBuilder,
-                          title: String,
-                          items: &[crate::output::report_model::ActionItem],
-                          i18n: &I18n|
-     -> renderreport::engine::ReportBuilder {
-        if items.is_empty() {
-            return b;
-        }
-        b = b.add_component(Section::new(title).with_level(2));
-        let mut table = AuditTable::new(vec![
-            TableColumn::new(i18n.t("column-action")),
-            TableColumn::new(effort_col),
-            TableColumn::new("Scope"),
-            TableColumn::new(role_col),
-            TableColumn::new(i18n.t("column-priority")),
-        ]);
-        for item in items {
-            let action_lower = item.action.to_lowercase();
-            let scope = if action_lower.contains("alle")
-                || action_lower.contains("global")
-                || action_lower.contains("designsystem")
-                || action_lower.contains("design system")
-                || action_lower.contains("seitenübergreifend")
-                || action_lower.contains("site-wide")
-            {
-                scope_global
-            } else if action_lower.contains("content")
-                || action_lower.contains("text")
-                || action_lower.contains("bild")
-                || action_lower.contains("image")
-            {
-                scope_content
-            } else {
-                scope_component
-            };
-            table = table.add_row(vec![
-                item.action.clone(),
-                item.effort.label().to_string(),
-                scope.to_string(),
-                role_label_i18n(item.role, i18n),
-                priority_label_i18n(item.priority, i18n),
-            ]);
-        }
-        b.add_component(table)
-    };
-
-    builder = render_section(
-        builder,
-        i18n.t("section-quick-wins"),
-        &plan.quick_wins,
-        i18n,
-    );
-    builder = render_section(
-        builder,
-        i18n.t("section-medium-actions"),
-        &plan.medium_term,
-        i18n,
-    );
-    builder = render_section(
-        builder,
-        i18n.t("section-structural-actions"),
-        &plan.structural,
-        i18n,
-    );
-    builder
-}
-
-/// Render cross-page consistency analysis (issues #44/#45/#46).
-/// Shows navigation, heading, and canonical consistency stats with any
-/// findings as warning rows.
-fn render_batch_consistency(
-    mut builder: renderreport::engine::ReportBuilder,
-    consistency: &crate::audit::batch_consistency::BatchConsistencyAnalysis,
-    i18n: &I18n,
-) -> renderreport::engine::ReportBuilder {
-    let en = i18n.locale() == "en";
-    let title = if en {
-        "Cross-page consistency"
-    } else {
-        "Seitenübergreifende Konsistenz"
-    };
-    let intro = if en {
-        "Checks that shared structural elements (navigation, headings, canonical URLs) are consistent across all audited pages. WCAG 3.2.3 / 3.2.4."
-    } else {
-        "Prüft, ob geteilte strukturelle Elemente (Navigation, Überschriften, Canonical-URLs) auf allen geprüften Seiten konsistent sind. WCAG 3.2.3 / 3.2.4."
-    };
-    builder = builder.add_component(SectionHeaderSplit::new(title, intro).with_level(2));
-
-    let nav = &consistency.navigation;
-    let nav_title = if en {
-        format!(
-            "Navigation ({}/{} with main nav, {}/{} with skip link)",
-            nav.pages_with_main_nav, nav.total_pages, nav.pages_with_skip_link, nav.total_pages
-        )
-    } else {
-        format!(
-            "Navigation ({}/{} mit Hauptnav, {}/{} mit Skip-Link)",
-            nav.pages_with_main_nav, nav.total_pages, nav.pages_with_skip_link, nav.total_pages
-        )
-    };
-    let nav_rows: Vec<ChecklistRow> = if nav.findings.is_empty() {
-        vec![ChecklistRow::new(
-            if en { "Consistent" } else { "Konsistent" },
-            if en {
-                "No navigation inconsistencies detected."
-            } else {
-                "Keine Navigation-Inkonsistenzen erkannt."
-            },
-        )
-        .with_status("good")]
-    } else {
-        nav.findings
-            .iter()
-            .map(|f| {
-                ChecklistRow::new(if en { "Issue" } else { "Befund" }, f.as_str())
-                    .with_status("warn")
-            })
-            .collect()
-    };
-    builder = builder.add_component(ChecklistPanel::new(nav_rows).with_title(&nav_title));
-
-    let h = &consistency.headings;
-    let head_title = if en {
-        format!(
-            "Headings ({}/{} with single H1, {} missing, {} multiple)",
-            h.pages_with_single_h1, h.total_pages, h.pages_with_no_h1, h.pages_with_multiple_h1
-        )
-    } else {
-        format!(
-            "Überschriften ({}/{} mit einem H1, {} ohne H1, {} mit mehreren)",
-            h.pages_with_single_h1, h.total_pages, h.pages_with_no_h1, h.pages_with_multiple_h1
-        )
-    };
-    let head_rows: Vec<ChecklistRow> = if h.findings.is_empty() {
-        vec![ChecklistRow::new(
-            if en { "Consistent" } else { "Konsistent" },
-            if en {
-                "Every page starts with exactly one H1."
-            } else {
-                "Jede Seite beginnt mit genau einem H1."
-            },
-        )
-        .with_status("good")]
-    } else {
-        h.findings
-            .iter()
-            .map(|f| {
-                ChecklistRow::new(if en { "Issue" } else { "Befund" }, f.as_str())
-                    .with_status("warn")
-            })
-            .collect()
-    };
-    builder = builder.add_component(ChecklistPanel::new(head_rows).with_title(&head_title));
-
-    let c = &consistency.canonical;
-    let canon_title = if en {
-        format!(
-            "Canonical URLs (www: {}, non-www: {}, missing: {} of {})",
-            c.www_count, c.non_www_count, c.missing_count, c.total_pages
-        )
-    } else {
-        format!(
-            "Canonical-URLs (www: {}, ohne www: {}, fehlend: {} von {})",
-            c.www_count, c.non_www_count, c.missing_count, c.total_pages
-        )
-    };
-    let canon_rows: Vec<ChecklistRow> = if c.findings.is_empty() {
-        vec![ChecklistRow::new(
-            if en { "Consistent" } else { "Konsistent" },
-            if en {
-                "All pages canonicalize to the same domain variant."
-            } else {
-                "Alle Seiten kanonisieren auf dieselbe Domain-Variante."
-            },
-        )
-        .with_status("good")]
-    } else {
-        c.findings
-            .iter()
-            .map(|f| {
-                ChecklistRow::new(if en { "Issue" } else { "Befund" }, f.as_str())
-                    .with_status("warn")
-            })
-            .collect()
-    };
-    builder.add_component(ChecklistPanel::new(canon_rows).with_title(&canon_title))
-}
-
-/// Closing section for batch report
-fn render_next_steps_batch(
-    mut builder: renderreport::engine::ReportBuilder,
-    pres: &BatchPresentation,
-    i18n: &I18n,
-) -> renderreport::engine::ReportBuilder {
-    let en = i18n.locale() == "en";
-    let intro = if en {
-        "Concrete recommendation for implementation."
-    } else {
-        "Konkrete Handlungsempfehlung für die Umsetzung."
-    };
-    builder = builder.add_component(
-        SectionHeaderSplit::new(i18n.t("section-next-steps-recommended"), intro).with_level(1),
-    );
-
-    let mut steps: Vec<(String, &str, &str)> = Vec::new();
-
-    let timeframe_quick = if en { "Week 1" } else { "Woche 1" };
-    let timeframe_medium = if en { "Week 2–3" } else { "Woche 2–3" };
-    let timeframe_structural = if en { "Month 1–2" } else { "Monat 1–2" };
-    let timeframe_medium_alt = if en { "Month 1" } else { "Monat 1" };
-    let scope_global = "global";
-    let scope_component = if en {
-        "component-based"
-    } else {
-        "komponentenbasiert"
-    };
-
-    // From quick wins
-    for item in &pres.action_plan.quick_wins {
-        if steps.len() >= 3 {
-            break;
-        }
-        let timeframe = match item.effort {
-            Effort::Quick => timeframe_quick,
-            Effort::Medium => timeframe_medium,
-            Effort::Structural => timeframe_structural,
-        };
-        let action_lower = item.action.to_lowercase();
-        let scope = if action_lower.contains("alle")
-            || action_lower.contains("designsystem")
-            || action_lower.contains("design system")
-            || action_lower.contains("global")
-        {
-            scope_global
-        } else {
-            scope_component
-        };
-        steps.push((item.action.clone(), timeframe, scope));
-    }
-
-    // Fallback from medium_term
-    if steps.len() < 3 {
-        for item in &pres.action_plan.medium_term {
-            if steps.len() >= 3 {
-                break;
-            }
-            steps.push((item.action.clone(), timeframe_medium_alt, scope_component));
-        }
-    }
-
-    if !steps.is_empty() {
-        let mut table = AuditTable::new(vec![
-            TableColumn::new(i18n.t("column-priority")),
-            TableColumn::new(i18n.t("column-action")),
-            TableColumn::new(i18n.t("column-timeframe")),
-            TableColumn::new("Scope"),
-        ]);
-        for (i, (action, timeframe, scope)) in steps.iter().enumerate() {
-            table = table.add_row(vec![
-                format!("{}", i + 1),
-                action.clone(),
-                timeframe.to_string(),
-                scope.to_string(),
-            ]);
-        }
-        builder = builder.add_component(table);
-    }
-
-    let callout_body = if en {
-        "For a complete WCAG conformance check we additionally recommend a manual audit with assistive technologies (screen reader, keyboard navigation). This automated audit covers about 30–40% of WCAG criteria."
-    } else {
-        "Für eine vollständige WCAG-Konformitätsprüfung empfehlen wir ergänzend einen manuellen Audit mit assistiven Technologien (Screenreader, Tastaturnavigation). Dieser automatisierte Audit deckt ca. 30–40% der WCAG-Kriterien ab."
-    };
-    builder = builder
-        .add_component(Callout::info(callout_body).with_title(i18n.t("section-next-steps-block")));
-
-    builder
-}
-
-// ─── Batch Report ───────────────────────────────────────────────────────────
-
-pub fn generate_batch_pdf(batch: &BatchReport, config: &ReportConfig) -> anyhow::Result<Vec<u8>> {
-    let engine = create_engine()?;
-    let i18n = I18n::new(&config.locale)?;
-    let pres = build_batch_presentation(batch);
-
-    let domain = &pres.portfolio_summary.domain;
-    let score = pres.portfolio_summary.average_score.round() as u32;
-
-    let mut builder = engine
-        .report("wcag-batch-audit")
-        .title(&pres.cover.title)
-        .subtitle(&pres.cover.url)
-        .metadata("date", &pres.cover.date)
-        .metadata("version", &pres.cover.version)
-        .metadata("author", domain)
-        .metadata("footer_prefix", "Audit:")
-        .metadata("footer_link_url", "")
-        .metadata(
-            "footer_tagline",
-            "A technical auditing platform by casoon.de",
-        );
-
-    let cover_logo_asset = cover_logo_asset(config);
-    builder = register_cover_logo_asset(builder, config, cover_logo_asset);
-
-    // ── Cover Page with Audit-Rahmen ────────────────────────────────
-    builder = builder
-        .add_component(Image::new(cover_logo_asset).with_width("120pt"))
-        .add_component(
-            Label::new(i18n.t("batch-cover-eyebrow"))
-                .with_size("11pt")
-                .bold()
-                .with_color("#0f766e"),
-        )
-        .add_component(
-            Label::new(i18n.t("batch-cover-title"))
-                .with_size("28pt")
-                .bold(),
-        )
-        .add_component(
-            Label::new(i18n.t("batch-cover-kicker"))
-                .with_size("12pt")
-                .with_color("#475569"),
-        );
-
-    // Audit-Rahmen box (matching single report style)
-    {
-        let modules_str = pres.portfolio_summary.active_modules.join(", ");
-
-        let mut cover_meta = KeyValueList::new().with_title(i18n.t("batch-cover-frame-title"));
-        cover_meta = cover_meta
-            .add(i18n.t("batch-cover-frame-domain"), domain)
-            .add(i18n.t("batch-cover-frame-date"), &pres.cover.date)
-            .add(
-                i18n.t("batch-cover-frame-urls"),
-                format!("{}", pres.portfolio_summary.total_urls),
-            )
-            .add(
-                i18n.t("batch-cover-frame-certificate"),
-                &pres.portfolio_summary.certificate,
-            )
-            .add(i18n.t("batch-cover-frame-modules"), &modules_str)
-            .add(
-                i18n.t("batch-cover-frame-version"),
-                format!("auditmysite v{}", pres.cover.version),
-            );
-        builder = builder.add_component(cover_meta);
-    }
-
-    let batch_badge_asset = "/certificate-badge-batch.svg";
-    let batch_badge_enabled =
-        if let Ok(path) = certificate_badge_path(batch_certificate_label(score)) {
-            builder = builder.asset(batch_badge_asset, path);
-            true
-        } else {
-            false
-        };
-
-    builder = builder
-        .add_component(build_batch_cover_score_row(
-            score,
-            pres.portfolio_summary.total_urls as u32,
-            pres.portfolio_summary.total_violations as u32,
-            batch_badge_enabled.then_some(batch_badge_asset),
-            &i18n,
-        )?)
-        .add_component(
-            TextBlock::new(&pres.portfolio_summary.verdict_text)
-                .with_size("11pt")
-                .with_line_height("1.4em")
-                .with_max_width("100%"),
-        )
-        .add_component(output_scope_callout(&i18n))
-        .add_component(PageBreak::new());
-    if config.level != ReportLevel::Executive {
-        builder = builder.add_component(TableOfContents::new().with_depth(1));
-    }
-
-    let dist = &pres.portfolio_summary.severity_distribution;
-
-    // ── 1. Status der Website ───────────────────────────────────────
-    {
-        // Block 1: Bewertung — klare Einordnung, Risiko primär
-        let assessment = build_batch_assessment(&pres.portfolio_summary, dist, &i18n);
-        let en_status = i18n.locale() == "en";
-        let risk_action = match (pres.portfolio_summary.risk_level.as_str(), en_status) {
-            ("Kritisch" | "Critical", true) => "Act immediately",
-            ("Kritisch" | "Critical", false) => "Sofort handeln",
-            ("Hoch" | "High", true) => "Fix soon",
-            ("Hoch" | "High", false) => "Zeitnah beheben",
-            ("Mittel" | "Medium", true) => "Improve deliberately",
-            ("Mittel" | "Medium", false) => "Gezielt verbessern",
-            (_, true) => "Maintain level",
-            (_, false) => "Niveau halten",
-        };
-        let risk_title = format!("{}  —  {}", assessment, risk_action);
-        let callout = match pres.portfolio_summary.risk_level.as_str() {
-            "Kritisch" | "Hoch" | "Critical" | "High" => {
-                Callout::warning(&pres.portfolio_summary.risk_summary).with_title(&risk_title)
-            }
-            "Mittel" | "Medium" => {
-                Callout::info(&pres.portfolio_summary.risk_summary).with_title(&risk_title)
-            }
-            _ => Callout::success(&pres.portfolio_summary.risk_summary).with_title(&risk_title),
-        };
-        builder = builder
-            .add_component(Section::new(i18n.t("batch-section-status")).with_level(1))
-            .add_component(callout);
-    }
-
-    // Score overview cards (sekundär)
-    builder = builder.add_component(build_batch_overview_grid(
-        pres.portfolio_summary.total_urls as u32,
-        score,
-        pres.portfolio_summary.total_violations as u32,
-        (dist.critical + dist.high) as u32,
-        pres.portfolio_summary.crawl_links.as_ref().map(|links| {
-            (links.broken_internal_links.len() + links.broken_external_links.len()) as u32
-        }),
-    ));
-
-    // Block 2: Kernaussagen (max 3 Punkte)
-    {
-        let key_points = build_batch_key_points(&pres, dist, &i18n);
-        let mut kp_list = List::new().with_title(i18n.t("narrative-key-points-title"));
-        for point in &key_points {
-            kp_list = kp_list.add_item(point);
-        }
-        builder = builder.add_component(kp_list);
-    }
-
-    // Auswirkungen
-    {
-        let en = i18n.locale() == "en";
-        let a11y_avg = pres.portfolio_summary.average_score.round() as u32;
-        let user_impact = match (a11y_avg, en) {
-            (s, true) if s < 50 => "Critical — core functions are unreachable for users with disabilities",
-            (s, false) if s < 50 => "Kritisch — zentrale Funktionen sind für Nutzer mit Einschränkungen nicht erreichbar",
-            (s, true) if s < 70 => "Limited — structural issues impede users with assistive technologies",
-            (s, false) if s < 70 => "Eingeschränkt — strukturelle Probleme behindern Nutzer mit Hilfstechnologien",
-            (s, true) if s < 85 => "Good — individual barriers for assistive technologies on several pages",
-            (s, false) if s < 85 => "Gut — einzelne Barrieren für Hilfstechnologien auf mehreren Seiten",
-            (_, true) => "Very good — assistive technologies are largely supported",
-            (_, false) => "Sehr gut — Hilfstechnologien werden weitgehend unterstützt",
-        };
-        let business_impact = if dist.critical > 0 {
-            if en {
-                "Large parts of the website are unusable or barely usable for certain user groups."
-            } else {
-                "Weite Teile der Website sind für bestimmte Nutzergruppen nicht oder kaum nutzbar."
-            }
-        } else if dist.high > 0 {
-            if en {
-                "Individual functional areas are problematic for users with disabilities."
-            } else {
-                "Einzelne Funktionsbereiche sind für Nutzer mit Einschränkungen problematisch."
-            }
-        } else if en {
-            "Low impact — users can fundamentally use the website."
-        } else {
-            "Geringe Auswirkung — Nutzer können die Website grundsätzlich verwenden."
-        };
-        let legal_impact = if dist.critical > 0 {
-            if en {
-                "WCAG Level A violations detected automatically — manual review required for a defensible BFSG classification."
-            } else {
-                "WCAG-Level-A-Verstöße automatisiert erkannt — für belastbare BFSG-Einordnung ist manuelle Prüfung nötig."
-            }
-        } else if en {
-            "No critical violations detected automatically — manual review recommended for full classification."
-        } else {
-            "Automatisiert keine kritischen Verstöße erkannt — manuelle Prüfung für vollständige Einordnung empfohlen."
-        };
-
-        let (user_label, business_label, risk_label) = if en {
-            ("User", "Business", "Risk")
-        } else {
-            ("Nutzer", "Business", "Risiko")
-        };
-        let mut impact_kv = KeyValueList::new().with_title(i18n.t("narrative-impact-title"));
-        impact_kv = impact_kv
-            .add(user_label, user_impact)
-            .add(business_label, business_impact)
-            .add(risk_label, legal_impact);
-        builder = builder.add_component(impact_kv);
-    }
-
-    // Block 3: Handlungsempfehlung
-    {
-        let actions = build_batch_quick_actions(&pres, &i18n);
-        if !actions.is_empty() {
-            let rows: Vec<ChecklistRow> = actions
-                .iter()
-                .map(|(action, _timeframe)| ChecklistRow::new(action, "").with_status("warn"))
-                .collect();
-            builder = builder.add_component(
-                ChecklistPanel::new(rows).with_title(i18n.t("narrative-quick-actions-title")),
-            );
-        }
-    }
-
-    // Module overview
-    if !pres.portfolio_summary.module_averages.is_empty() {
-        let mut module_kv = KeyValueList::new().with_title(i18n.t("batch-panel-module-averages"));
-        for (name, score) in &pres.portfolio_summary.module_averages {
-            module_kv = module_kv.add(name, format!("{}/100", score));
-        }
-        builder = builder.add_component(module_kv);
-    }
-
-    // ── 2. URL-Ranking ──────────────────────────────────────────────
-    let rows: Vec<BenchmarkRow> = pres
-        .url_ranking
-        .iter()
-        .enumerate()
-        .map(|(i, u)| {
-            let mut row = BenchmarkRow::new(
-                (i + 1) as u32,
-                &truncate_url(&u.url, 35),
-                u.overall_score,
-                u.score as u32,
-                u.critical_violations as u32,
-            );
-            if let Some(detail) = pres.url_details.iter().find(|detail| detail.url == u.url) {
-                if let Some((_, score)) = detail
-                    .module_scores
-                    .iter()
-                    .find(|(module, _)| module == "SEO")
-                {
-                    row = row.with_seo(*score);
-                }
-                if let Some((_, score)) = detail
-                    .module_scores
-                    .iter()
-                    .find(|(module, _)| module == "Performance")
-                {
-                    row = row.with_performance(*score);
-                }
-                if let Some((_, score)) = detail
-                    .module_scores
-                    .iter()
-                    .find(|(module, _)| module == "Security")
-                {
-                    row = row.with_security(*score);
-                }
-            }
-            row
-        })
-        .collect();
-
-    builder = builder
-        .add_component(
-            SectionHeaderSplit::new(
-                "URL-Ranking",
-                "Übersicht aller geprüften URLs, sortiert nach Score. \
-             URLs mit niedrigerem Score haben höheren Handlungsbedarf.",
-            )
-            .with_level(1),
-        )
-        .add_component(BenchmarkTable::new(rows));
-
-    // ── 3. Top-Probleme (vereinheitlicht) ─────────────────────────
-    let en_top = i18n.locale() == "en";
-    let top_intro = if en_top {
-        "The following problem groups occur across multiple URLs. \
-         Fixing them yields the largest improvement because they affect many pages simultaneously."
-    } else {
-        "Die folgenden Problemgruppen treten über mehrere URLs hinweg auf. \
-         Durch Behebung dieser Probleme wird die größte Verbesserung erzielt, \
-         da sie viele Seiten gleichzeitig betreffen."
-    };
-    builder = builder.add_component(
-        SectionHeaderSplit::new(i18n.t("batch-section-most-frequent"), top_intro).with_level(1),
-    );
-
-    // Übersichtstabelle mit Aufwand
-    if !pres.issue_frequency.is_empty() {
-        let affected_col = if en_top {
-            "Affected URLs"
-        } else {
-            "Betr. URLs"
-        };
-        let mut freq_table = AuditTable::new(vec![
-            TableColumn::new(i18n.t("batch-col-problem")),
-            TableColumn::new("WCAG"),
-            TableColumn::new(i18n.t("batch-col-occurrences")),
-            TableColumn::new(affected_col),
-            TableColumn::new(i18n.t("batch-col-priority")),
-        ])
-        .with_title(i18n.t("batch-section-most-frequent-violations"));
-
-        for issue in &pres.issue_frequency {
-            freq_table = freq_table.add_row(vec![
-                issue.problem.clone(),
-                issue.wcag.clone(),
-                issue.occurrences.to_string(),
-                issue.affected_urls.to_string(),
-                self::helpers::priority_label_i18n(issue.priority, &i18n),
-            ]);
-        }
-        builder = builder.add_component(freq_table);
-    }
-
-    // Unified problem blocks — 1 Problem = 1 kompakter Block
-    // (keine doppelten Cards + Details mehr)
-    let scope_global_word = if en_top {
-        "global (all pages)"
-    } else {
-        "global (alle Seiten)"
-    };
-    let scope_individual = if en_top {
-        "individual pages"
-    } else {
-        "einzelne Seiten"
-    };
-    let occurrences_word_top = if en_top { "occurrences" } else { "Vorkommen" };
-    let affected_urls_word = if en_top {
-        "affected URLs"
-    } else {
-        "betroffene URLs"
-    };
-    let effort_word = if en_top { "Effort" } else { "Aufwand" };
-    let scope_word = "Scope";
-    let impact_user_label = if en_top {
-        "Impact (user)"
-    } else {
-        "Impact (Nutzer)"
-    };
-    let impact_business_label = if en_top {
-        "Impact (business)"
-    } else {
-        "Impact (Business)"
-    };
-    let fix_label = "Fix";
-    let meta_label = "Meta";
-
-    for group in pres.top_issues.iter().take(5) {
-        let scope = if group.affected_urls.len() >= pres.portfolio_summary.total_urls {
-            scope_global_word
-        } else {
-            scope_individual
-        };
-        let effort_label = group.effort.label();
-        let meta_line = format!(
-            "{} {} · {} {} · {}: {} · {}: {}",
-            group.occurrence_count,
-            occurrences_word_top,
-            group.affected_urls.len(),
-            affected_urls_word,
-            effort_word,
-            effort_label,
-            scope_word,
-            scope
-        );
-
-        let mut kv = KeyValueList::new().with_title(&group.title);
-        kv = kv
-            .add(
-                i18n.t("findings-card-key-problem"),
-                &group.customer_description,
-            )
-            .add(impact_user_label, &group.user_impact)
-            .add(impact_business_label, &group.business_impact)
-            .add(i18n.t("findings-card-key-cause"), &group.typical_cause)
-            .add(fix_label, &group.recommendation)
-            .add(meta_label, meta_line);
-        builder = builder.add_component(kv);
-    }
-
-    // ── 4. Maßnahmenplan (mit Aufwand + Scope) ─────────────────────
-    let action_intro = if en_top {
-        "The following actions are prioritized by effort and impact. \
-         Actions that improve many pages simultaneously take precedence."
-    } else {
-        "Die folgenden Maßnahmen sind nach Aufwand und Wirkung priorisiert. \
-         Maßnahmen, die viele Seiten gleichzeitig verbessern, haben Vorrang."
-    };
-    builder = builder.add_component(
-        SectionHeaderSplit::new(i18n.t("batch-action-plan-title"), action_intro).with_level(1),
-    );
-    builder = render_batch_action_plan_enhanced(builder, &pres.action_plan, &i18n);
-
-    // ── 5a. Render Blocking (Batch) ─────────────────────────────────
-    if !pres.portfolio_summary.render_blocking_summary.is_empty() {
-        let mut kv = KeyValueList::new().with_title("Render-Blocking-Übersicht (domainweit)");
-        for (label, value) in &pres.portfolio_summary.render_blocking_summary {
-            kv = kv.add(label, value);
-        }
-        builder = builder
-            .add_component(Section::new("Render-Blocking & Assets").with_level(1))
-            .add_component(TextBlock::new(
-                "Aggregierte Auswertung render-blockierender Ressourcen über alle geprüften Seiten. \
-                 Blocking-Scripts und -CSS verzögern den First Contentful Paint. \
-                 Third-Party-Traffic entsteht durch externe Fonts, Analytics und Widgets.",
-            ))
-            .add_component(kv);
-    }
-
-    // ── 5b. Performance Budgets (Batch) ─────────────────────────────
-    let en_pdf = i18n.locale() == "en";
-    if !pres.portfolio_summary.budget_summary.is_empty() {
-        let pages_col = if en_pdf {
-            "Affected pages"
-        } else {
-            "Betr. Seiten"
-        };
-        let budget_table_title = if en_pdf {
-            "Performance budget violations (domain-wide)"
-        } else {
-            "Performance-Budget-Verstöße (domainweit)"
-        };
-        let mut table = AuditTable::new(vec![
-            TableColumn::new(i18n.t("batch-col-metric")),
-            TableColumn::new(i18n.t("batch-col-budget")),
-            TableColumn::new(pages_col),
-            TableColumn::new(i18n.t("batch-col-severity")),
-        ])
-        .with_title(budget_table_title);
-        for (metric, budget, count, sev) in &pres.portfolio_summary.budget_summary {
-            table = table.add_row(vec![
-                metric.clone(),
-                budget.clone(),
-                count.to_string(),
-                sev.clone(),
-            ]);
-        }
-        let budgets_intro = if en_pdf {
-            "Performance budgets define limits for load times, asset sizes and third-party \
-             traffic. The following table shows on how many pages each budget was exceeded."
-        } else {
-            "Performance-Budgets definieren Obergrenzen für Ladezeiten, Asset-Größen und \
-             Drittanbieter-Traffic. Die folgende Tabelle zeigt, auf wie vielen Seiten welche \
-             Budgets überschritten wurden."
-        };
-        builder = builder
-            .add_component(Section::new(i18n.t("batch-section-performance-budgets")).with_level(1))
-            .add_component(TextBlock::new(budgets_intro))
-            .add_component(table);
-    }
-
-    // ── 5. Technische URL-Matrix ───────────────────────────────────
-    builder = builder.add_component(
-        SectionHeaderSplit::new(
-            i18n.t("batch-section-tech-url-matrix"),
-            i18n.t("batch-section-tech-url-matrix-intro"),
-        )
-        .with_level(1),
-    );
-
-    if let Some(ref crawl_links) = pres.portfolio_summary.crawl_links {
-        let target_col = if en_pdf { "Target" } else { "Ziel" };
-        let type_col = if en_pdf { "Type" } else { "Typ" };
-        let direct_label = if en_pdf { "direct" } else { "direkt" };
-        let hops_label = if en_pdf { "hops" } else { "Hops" };
-        let internal_intro = if en_pdf {
-            format!(
-                "For the crawl starting at {} we checked {} internal link targets. {} broken internal links detected.",
-                crawl_links.seed_url,
-                crawl_links.checked_internal_links,
-                crawl_links.broken_internal_links.len()
-            )
-        } else {
-            format!(
-                "Für den Crawl ab {} wurden {} interne Linkziele geprüft. {} kaputte interne Verlinkungen wurden erkannt.",
-                crawl_links.seed_url,
-                crawl_links.checked_internal_links,
-                crawl_links.broken_internal_links.len()
-            )
-        };
-        builder = builder
-            .add_component(
-                Section::new(i18n.t("batch-section-broken-links-internal")).with_level(1),
-            )
-            .add_component(TextBlock::new(internal_intro));
-
-        if crawl_links.broken_internal_links.is_empty() {
-            builder = builder.add_component(Callout::info(if en_pdf {
-                "No broken internal links detected in the audited crawl set."
-            } else {
-                "Keine kaputten internen Links im geprüften Crawl-Set erkannt."
-            }));
-        } else {
-            let mut table = AuditTable::new(vec![
-                TableColumn::new(i18n.t("batch-col-source")),
-                TableColumn::new(target_col),
-                TableColumn::new(i18n.t("batch-col-status-code")),
-                TableColumn::new(type_col),
-            ])
-            .with_title(i18n.t("batch-table-broken-internal"));
-
-            for row in &crawl_links.broken_internal_links {
-                let severity_color = match row.severity.as_str() {
-                    "high" => "#dc2626",
-                    "medium" => "#ea580c",
-                    _ => "#ca8a04",
-                };
-                let typ_label = if row.redirect_hops > 0 {
-                    format!("→{} {}", row.redirect_hops, hops_label)
-                } else {
-                    direct_label.to_string()
-                };
-                table = table.add_row(vec![
-                    truncate_url(&row.source_url, 30),
-                    truncate_url(&row.target_url, 38),
-                    format!("\x1b[{}m{}\x1b[0m", severity_color, row.status),
-                    typ_label,
-                ]);
-            }
-
-            builder = builder.add_component(table);
-        }
-
-        // External broken links
-        if !crawl_links.broken_external_links.is_empty() {
-            let ext_intro = if en_pdf {
-                format!(
-                    "{} external link targets checked. {} broken external links detected.",
-                    crawl_links.checked_external_links,
-                    crawl_links.broken_external_links.len()
-                )
-            } else {
-                format!(
-                    "{} externe Linkziele geprüft. {} kaputte externe Verlinkungen erkannt.",
-                    crawl_links.checked_external_links,
-                    crawl_links.broken_external_links.len()
-                )
-            };
-            builder = builder
-                .add_component(
-                    Section::new(i18n.t("batch-section-broken-links-external")).with_level(2),
-                )
-                .add_component(TextBlock::new(ext_intro));
-
-            let mut ext_table = AuditTable::new(vec![
-                TableColumn::new(i18n.t("batch-col-source")),
-                TableColumn::new(target_col),
-                TableColumn::new(i18n.t("batch-col-status-code")),
-                TableColumn::new(type_col),
-            ])
-            .with_title(i18n.t("batch-table-broken-external"));
-
-            for row in &crawl_links.broken_external_links {
-                let typ_label = if row.redirect_hops > 0 {
-                    format!("→{} {}", row.redirect_hops, hops_label)
-                } else {
-                    direct_label.to_string()
-                };
-                ext_table = ext_table.add_row(vec![
-                    truncate_url(&row.source_url, 30),
-                    truncate_url(&row.target_url, 38),
-                    row.status.clone(),
-                    typ_label,
-                ]);
-            }
-
-            builder = builder.add_component(ext_table);
-        } else if crawl_links.checked_external_links > 0 {
-            let ext_clean_msg = if en_pdf {
-                format!(
-                    "{} external link targets checked — no broken external links detected.",
-                    crawl_links.checked_external_links
-                )
-            } else {
-                format!(
-                    "{} externe Linkziele geprüft — keine kaputten externen Links erkannt.",
-                    crawl_links.checked_external_links
-                )
-            };
-            builder = builder
-                .add_component(Section::new(i18n.t("batch-section-external-links")).with_level(2))
-                .add_component(Callout::info(ext_clean_msg));
-        }
-
-        // Redirect chains
-        if !crawl_links.redirect_chains.is_empty() {
-            let chain_intro = if en_pdf {
-                format!(
-                    "{} links with more than one redirect hop detected.",
-                    crawl_links.redirect_chains.len()
-                )
-            } else {
-                format!(
-                    "{} Links mit mehr als einem Redirect-Hop erkannt.",
-                    crawl_links.redirect_chains.len()
-                )
-            };
-            builder = builder
-                .add_component(Section::new(i18n.t("batch-section-redirect-chains")).with_level(2))
-                .add_component(TextBlock::new(chain_intro));
-
-            let mut chain_table = AuditTable::new(vec![
-                TableColumn::new(i18n.t("batch-col-source")),
-                TableColumn::new("Ziel"),
-                TableColumn::new("Hops"),
-                TableColumn::new("Final-URL"),
-            ])
-            .with_title("Redirect-Ketten (> 1 Hop)");
-
-            for chain in &crawl_links.redirect_chains {
-                chain_table = chain_table.add_row(vec![
-                    truncate_url(&chain.source_url, 28),
-                    truncate_url(&chain.target_url, 28),
-                    chain.hops.to_string(),
-                    truncate_url(&chain.final_url, 32),
-                ]);
-            }
-
-            builder = builder.add_component(chain_table);
-        }
-    }
-
-    let page_col = if en_pdf { "Page" } else { "Seite" };
-    let title_col = if en_pdf { "Title" } else { "Titel" };
-    let mut matrix = AuditTable::new(vec![
-        TableColumn::new("#").with_width("4%"),
-        TableColumn::new(page_col).with_width("26%"),
-        TableColumn::new(title_col).with_width("28%"),
-        TableColumn::new(i18n.t("batch-col-links-to")).with_width("10%"),
-        TableColumn::new(i18n.t("batch-col-links-from")).with_width("10%"),
-        TableColumn::new(i18n.t("batch-col-words")).with_width("10%"),
-        TableColumn::new("Score").with_width("12%"),
-    ])
-    .with_title(i18n.t("batch-table-pages-overview"));
-
-    for row in &pres.url_matrix {
-        let score_str = pres
-            .url_details
-            .iter()
-            .find(|d| d.url == row.url)
-            .map(|d| format!("{}/100", d.score.round() as u32))
-            .unwrap_or_else(|| "—".to_string());
-        matrix = matrix.add_row(vec![
-            row.rank.to_string(),
-            truncate_url(&row.url, 34),
-            row.title
-                .as_deref()
-                .map(|t| truncate_url(t, 36))
-                .unwrap_or_else(|| "—".to_string()),
-            row.inbound_links.to_string(),
-            row.outbound_links.to_string(),
-            format_word_count(row.word_count),
-            score_str,
-        ]);
-    }
-    builder = builder.add_component(matrix);
-
-    if config.level != ReportLevel::Executive {
-        let mut focus_table = AuditTable::new(vec![
-            TableColumn::new("URL"),
-            TableColumn::new(i18n.t("batch-col-page-type")),
-            TableColumn::new(i18n.t("batch-col-attributes")),
-            TableColumn::new(i18n.t("batch-col-top-issues")),
-        ])
-        .with_title(i18n.t("batch-table-focus-pages"));
-
-        for detail in pres.url_details.iter().take(10) {
-            focus_table = focus_table.add_row(vec![
-                truncate_url(&detail.url, 38),
-                detail.page_type.clone().unwrap_or_else(|| "—".to_string()),
-                if detail.page_attributes.is_empty() {
-                    "—".to_string()
-                } else {
-                    truncate_url(&detail.page_attributes.join(", "), 40)
-                },
-                if detail.top_issues.is_empty() {
-                    "—".to_string()
-                } else {
-                    truncate_url(&detail.top_issues.join(", "), 52)
-                },
-            ]);
-        }
-        builder = builder.add_component(focus_table);
-    }
-
-    // ── 6. Content & SEO — integriert mit Business-Impact ─────────
-    let en_batch = i18n.locale() == "en";
-    let (content_seo_title, content_seo_intro) = if en_batch {
-        (
-            "Content & SEO potential",
-            "Content strengths and weaknesses with direct relevance to rankings, visibility and conversion. \
-             Each finding is tied to a concrete action.",
-        )
-    } else {
-        (
-            "Content & SEO-Potenzial",
-            "Content-Stärken und -Schwächen mit direktem Bezug zu Rankings, Sichtbarkeit und Conversion. \
-             Jede Auffälligkeit ist an eine konkrete Handlung geknüpft.",
-        )
-    };
-    builder = builder
-        .add_component(SectionHeaderSplit::new(content_seo_title, content_seo_intro).with_level(1));
-
-    // Schwache Seiten zuerst — mit Business-Impact
-    if !pres.portfolio_summary.weakest_content_pages.is_empty() {
-        let issues_title = if en_batch {
-            "Content issues needing action"
-        } else {
-            "Content-Probleme mit Handlungsbedarf"
-        };
-        let high_marker = if en_batch { "high" } else { "hoch" };
-        let mut issues_kv = KeyValueList::new().with_title(issues_title);
-        for (url, page_type, score) in &pres.portfolio_summary.weakest_content_pages {
-            let relevance = business_relevance(Some(page_type.as_str()), url, i18n.locale());
-            let impact = if relevance == high_marker {
-                if en_batch {
-                    "Likely ranking loss + lower conversion"
-                } else {
-                    "Rankingverlust + geringere Conversion wahrscheinlich"
-                }
-            } else if *score < 30 {
-                if en_batch {
-                    "Weak organic visibility"
-                } else {
-                    "Schwache organische Sichtbarkeit"
-                }
-            } else if en_batch {
-                "Optimization potential for SEO"
-            } else {
-                "Optimierungspotenzial für SEO"
-            };
-            let value = if en_batch {
-                format!(
-                    "{} — {} → +300–800 words of structured content recommended",
-                    page_type, impact
-                )
-            } else {
-                format!(
-                    "{} — {} → +300–800 Wörter strukturierter Inhalt empfohlen",
-                    page_type, impact
-                )
-            };
-            let key_str = if en_batch {
-                format!("{} (profile: {}/100)", truncate_url(url, 35), score)
-            } else {
-                format!("{} (Profil: {}/100)", truncate_url(url, 35), score)
-            };
-            issues_kv = issues_kv.add(key_str, value);
-        }
-        builder = builder.add_component(issues_kv);
-    }
-
-    // Content-Auffälligkeiten als Business-Relevanz
-    if !pres.portfolio_summary.distribution_insights.is_empty() {
-        let action_label = if en_batch {
-            "Action needed"
-        } else {
-            "Handlungsbedarf"
-        };
-        let panel_title = if en_batch {
-            "Content patterns → business impact"
-        } else {
-            "Content-Auffälligkeiten → Business-Impact"
-        };
-        let rows: Vec<ChecklistRow> = pres
-            .portfolio_summary
-            .distribution_insights
-            .iter()
-            .map(|insight| {
-                let impact = if insight.contains("Thin") || insight.contains("dünn") {
-                    if en_batch {
-                        format!("{} → weaker rankings, lower dwell time", insight)
-                    } else {
-                        format!("{} → schwächere Rankings, geringere Verweildauer", insight)
-                    }
-                } else if insight.contains("Duplikat") || insight.contains("duplicate") {
-                    if en_batch {
-                        format!(
-                            "{} → keyword cannibalization, split ranking signals",
-                            insight
-                        )
-                    } else {
-                        format!(
-                            "{} → Keyword-Kannibalisierung, Split der Ranking-Signale",
-                            insight
-                        )
-                    }
-                } else {
-                    insight.clone()
-                };
-                ChecklistRow::new(action_label, &impact).with_status("warn")
-            })
-            .collect();
-        builder = builder.add_component(ChecklistPanel::new(rows).with_title(panel_title));
-    }
-
-    // Near-duplicates mit Business-Kontext
-    if !pres.portfolio_summary.near_duplicates.is_empty() {
-        let near_dup_title = if en_batch {
-            "Near-duplicate content → keyword cannibalization"
-        } else {
-            "Near-Duplicate-Content → Keyword-Kannibalisierung"
-        };
-        let mut table = AuditTable::new(vec![
-            TableColumn::new(i18n.t("batch-col-page-a")),
-            TableColumn::new(i18n.t("batch-col-page-b")),
-            TableColumn::new(i18n.t("batch-col-similarity")),
-            TableColumn::new(i18n.t("batch-col-risk")),
-        ])
-        .with_title(near_dup_title);
-
-        for (url_a, url_b, sim) in &pres.portfolio_summary.near_duplicates {
-            let risk = if *sim >= 95 {
-                if en_batch {
-                    "High — consolidate"
-                } else {
-                    "Hoch — konsolidieren"
-                }
-            } else if en_batch {
-                "Medium — differentiate"
-            } else {
-                "Mittel — differenzieren"
-            };
-            table = table.add_row(vec![
-                truncate_url(url_a, 35),
-                truncate_url(url_b, 35),
-                format!("{sim}%"),
-                risk.to_string(),
-            ]);
-        }
-        builder = builder.add_component(table);
-    }
-
-    // Seitentyp-Verteilung (kompakt)
-    if !pres.portfolio_summary.page_type_distribution.is_empty() {
-        let high_label = if en_batch { "high" } else { "hoch" };
-        let medium_label = if en_batch { "medium" } else { "mittel" };
-        let low_label = if en_batch { "low" } else { "niedrig" };
-        let mut type_table = AuditTable::new(vec![
-            TableColumn::new(i18n.t("batch-col-page-type")),
-            TableColumn::new(i18n.t("batch-col-pages-list")),
-            TableColumn::new(i18n.t("batch-col-share")),
-            TableColumn::new(i18n.t("batch-col-relevance")),
-        ])
-        .with_title(i18n.t("batch-table-page-type-distribution"));
-
-        for (label, count, pct) in &pres.portfolio_summary.page_type_distribution {
-            let relevance = match label.as_str() {
-                "Marketing / Landing Page"
-                | "Transaktional / Utility"
-                | "Transactional / Utility" => high_label,
-                "Editorial / Artikel"
-                | "Editorial / Article"
-                | "Strukturierter Wissensinhalt"
-                | "Structured knowledge content" => medium_label,
-                "Thin / Minimal Content" => low_label,
-                _ => medium_label,
-            };
-            type_table = type_table.add_row(vec![
-                label.clone(),
-                count.to_string(),
-                format!("{pct}%"),
-                relevance.to_string(),
-            ]);
-        }
-        builder = builder.add_component(type_table);
-    }
-
-    // Schema-Typ-Verteilung
-    if !pres.portfolio_summary.schema_distribution.is_empty() {
-        let total = pres.portfolio_summary.total_urls;
-        let without = pres.portfolio_summary.pages_without_schema;
-        let summary = if without == 0 {
-            if en_batch {
-                format!("All {} pages have structured data.", total)
-            } else {
-                format!("Alle {} Seiten haben strukturierte Daten.", total)
-            }
-        } else if en_batch {
-            format!("{} of {} pages without structured data.", without, total)
-        } else {
-            format!("{} von {} Seiten ohne strukturierte Daten.", without, total)
-        };
-        let schema_callout_title = if en_batch {
-            "Structured data (Schema.org)"
-        } else {
-            "Strukturierte Daten (Schema.org)"
-        };
-        builder = builder.add_component(Callout::info(&summary).with_title(schema_callout_title));
-        let mut schema_table = AuditTable::new(vec![
-            TableColumn::new(i18n.t("batch-col-schema-type")).with_width("55%"),
-            TableColumn::new(i18n.t("batch-col-pages-list")).with_width("20%"),
-            TableColumn::new(i18n.t("batch-col-share")).with_width("25%"),
-        ])
-        .with_title(i18n.t("batch-table-schema-distribution"));
-        for (schema_type, count) in &pres.portfolio_summary.schema_distribution {
-            let pct = (*count * 100).checked_div(total).unwrap_or(0);
-            schema_table = schema_table.add_row(vec![
-                schema_type.clone(),
-                count.to_string(),
-                format!("{pct}%"),
-            ]);
-        }
-        builder = builder.add_component(schema_table);
-    }
-
-    // Stärkste Seiten (kurz)
-    if !pres.portfolio_summary.strongest_content_pages.is_empty() {
-        let mut strengths = AuditTable::new(vec![
-            TableColumn::new("URL"),
-            TableColumn::new(i18n.t("batch-col-page-type")),
-            TableColumn::new(i18n.t("batch-col-profile")),
-        ])
-        .with_title(i18n.t("batch-table-top-pages"));
-
-        for (url, page_type, score) in &pres.portfolio_summary.strongest_content_pages {
-            strengths = strengths.add_row(vec![
-                truncate_url(url, 42),
-                page_type.clone(),
-                format!("{score}/100"),
-            ]);
-        }
-        builder = builder.add_component(strengths);
-    }
-
-    // ── Cross-page consistency (issues #44/#45/#46) ─────────────────
-    if let Some(ref consistency) = batch.consistency {
-        builder = render_batch_consistency(builder, consistency, &i18n);
-    }
-
-    // ── Empfohlene nächste Schritte ───────────────────────────────
-    builder = render_next_steps_batch(builder, &pres, &i18n);
-
-    // ── 7. Anhang ───────────────────────────────────────────────────
-    if config.level == ReportLevel::Technical && !pres.appendix.per_url.is_empty() {
-        let appendix_intro = if en_pdf {
-            "Complete listing of all detected violations per URL with technical details for implementation."
-        } else {
-            "Vollständige Auflistung aller erkannten Verstöße pro URL \
-             mit technischen Details für die Umsetzung."
-        };
-        builder = builder.add_component(
-            SectionHeaderSplit::new(i18n.t("section-appendix"), appendix_intro).with_level(1),
-        );
-
-        let rule_col = if en_pdf { "Rule" } else { "Regel" };
-        let elements_col = if en_pdf {
-            "Affected elements"
-        } else {
-            "Betr. Elemente"
-        };
-        for url_appendix in &pres.appendix.per_url {
-            if url_appendix.violations.is_empty() {
-                continue;
-            }
-
-            builder = builder
-                .add_component(Section::new(truncate_url(&url_appendix.url, 70)).with_level(2));
-
-            let mut table = AuditTable::new(vec![
-                TableColumn::new(rule_col),
-                TableColumn::new(i18n.t("batch-col-severity")),
-                TableColumn::new(i18n.t("batch-col-description")),
-                TableColumn::new(elements_col),
-            ]);
-
-            for v in &url_appendix.violations {
-                let elements = v
-                    .affected_elements
-                    .iter()
-                    .map(|e| e.selector.clone())
-                    .collect::<Vec<_>>()
-                    .join("; ");
-                table = table.add_row(vec![
-                    format!(
-                        "{} — {} ({}×)",
-                        v.rule,
-                        v.rule_name,
-                        v.affected_elements.len()
-                    ),
-                    severity_label_i18n(v.severity, &i18n),
-                    v.message.clone(),
-                    elements,
-                ]);
-            }
-            builder = builder.add_component(table);
-        }
-    }
-
-    let built_report = builder.build();
-    Ok(engine.render_pdf(&built_report)?)
-}
-
-// ─── Comparison Report ───────────────────────────────────────────────────────
-
-/// Generate a competitive comparison PDF report.
-pub fn generate_comparison_pdf(
-    comparison: &crate::audit::ComparisonReport,
-    config: &ReportConfig,
-) -> anyhow::Result<Vec<u8>> {
-    let engine = create_engine()?;
-
-    let avg_score = if comparison.entries.is_empty() {
-        0u32
-    } else {
-        (comparison
-            .entries
-            .iter()
-            .map(|e| e.overall_score as u64)
-            .sum::<u64>()
-            / comparison.entries.len() as u64) as u32
-    };
-
-    let author = comparison
-        .entries
-        .first()
-        .map(|e| extract_domain(&e.url))
-        .unwrap_or_default();
-
-    let i18n = I18n::new(&config.locale)?;
-    let en_cmp = i18n.locale() == "en";
-    let date_format = if en_cmp { "%Y-%m-%d" } else { "%d.%m.%Y" };
-    let mut builder = engine
-        .report("wcag-comparison")
-        .metadata("date", chrono::Local::now().format(date_format).to_string())
-        .metadata("author", &author)
-        .metadata("footer_prefix", "Audit:")
-        .metadata("footer_link_url", "")
-        .metadata(
-            "footer_tagline",
-            "A technical auditing platform by casoon.de",
-        );
-
-    let cover_logo_asset = cover_logo_asset(config);
-    builder = register_cover_logo_asset(builder, config, cover_logo_asset);
-
-    let comparison_subline = if en_cmp {
-        format!(
-            "Comparison of {} domains — avg overall: {}/100",
-            comparison.entries.len(),
-            avg_score
-        )
-    } else {
-        format!(
-            "Vergleich von {} Domains — Ø Gesamt: {}/100",
-            comparison.entries.len(),
-            avg_score
-        )
-    };
-
-    builder = builder
-        .add_component(Image::new(cover_logo_asset).with_width("22%"))
-        .add_component(
-            Label::new(&author)
-                .with_size("10pt")
-                .bold()
-                .with_color("#0f766e"),
-        )
-        .add_component(
-            Label::new(i18n.t("comparison-cover-title"))
-                .with_size("28pt")
-                .bold(),
-        )
-        .add_component(
-            Label::new(comparison_subline)
-                .with_size("12pt")
-                .with_color("#475569"),
-        )
-        .add_component(
-            Label::new(format!("auditmysite v{}", env!("CARGO_PKG_VERSION")))
-                .with_size("9pt")
-                .with_color("#94a3b8"),
-        )
-        .add_component(PageBreak::new());
-
-    // ── 1. Domain-Ranking ────────────────────────────────────────────
-    let ranking_intro = if en_cmp {
-        format!(
-            "Comparison of {} domains based on a full audit of each home page. \
-             Average overall score: {}/100.",
-            comparison.entries.len(),
-            avg_score,
-        )
-    } else {
-        format!(
-            "Vergleich von {} Domains anhand eines vollständigen Audits der jeweiligen Startseite. \
-             Durchschnittlicher Gesamtscore: {}/100.",
-            comparison.entries.len(),
-            avg_score,
-        )
-    };
-    builder = builder
-        .add_component(Section::new(i18n.t("comparison-domain-ranking")).with_level(1))
-        .add_component(TextBlock::new(ranking_intro));
-
-    let rows: Vec<BenchmarkRow> = comparison
-        .entries
-        .iter()
-        .enumerate()
-        .map(|(i, e)| {
-            let mut row = BenchmarkRow::new(
-                (i + 1) as u32,
-                &e.domain,
-                e.overall_score,
-                e.accessibility_score,
-                e.critical_violations as u32,
-            );
-            if let Some(s) = e.seo_score {
-                row = row.with_seo(s);
-            }
-            if let Some(s) = e.performance_score {
-                row = row.with_performance(s);
-            }
-            if let Some(s) = e.security_score {
-                row = row.with_security(s);
-            }
-            row
-        })
-        .collect();
-
-    builder = builder
-        .add_component(BenchmarkTable::new(rows).with_title(&i18n.t("comparison-domain-ranking")));
-
-    // ── 2. Modul-Vergleich ───────────────────────────────────────────
-    let has_module_data = comparison
-        .entries
-        .iter()
-        .any(|e| e.seo_score.is_some() || e.performance_score.is_some());
-
-    if has_module_data {
-        builder = builder
-            .add_component(Section::new(i18n.t("comparison-module-comparison")).with_level(1));
-
-        let comparison_modules: Vec<ComparisonModule> = comparison
-            .entries
-            .iter()
-            .map(|e| ComparisonModule::new(&e.domain, e.overall_score))
-            .collect();
-        builder = builder.add_component(ModuleComparison::new(comparison_modules));
-    }
-
-    // ── 3. Top Findings je Domain ────────────────────────────────────
-    let has_issues = comparison.entries.iter().any(|e| !e.top_issues.is_empty());
-    if has_issues {
-        builder = builder.add_component(
-            Section::new(i18n.t("comparison-top-findings-per-domain")).with_level(1),
-        );
-
-        let top_findings_label = if en_cmp {
-            "Top findings"
-        } else {
-            "Top Findings"
-        };
-        for entry in &comparison.entries {
-            if entry.top_issues.is_empty() {
-                continue;
-            }
-            builder = builder.add_component(
-                Section::new(format!("{} — {}", entry.domain, top_findings_label)).with_level(2),
-            );
-            let mut list = List::new();
-            for issue in &entry.top_issues {
-                list = list.add_item(issue.clone());
-            }
-            builder = builder.add_component(list);
-        }
-    }
-
-    let built_report = builder.build();
-    Ok(engine.render_pdf(&built_report)?)
 }
 
 #[cfg(all(test, feature = "pdf_test"))]
